pyproject.toml
[build-system]
    build-backend = "poetry.core.masonry.api"
    requires      = ["poetry-core>=1.0.0"]
[tool.poetry]
    authors = ["Tim Holzapfel"]
    classifiers = [
        "Environment :: Web Environment :: Mozilla",
        "Development Status :: 5 - Production/Stable",
    ]
    description = "Resize song coverart."
    license = "Proprietary"
    name = "ResizeImage"
    packages = [{ include = "ResizeImage", from = "src" }]
    version = "1.0.0"
[tool.poetry.dependencies]
    Pillow                = "^9.0.1"
    black                 = "^22.1.0"
    joblib                = "^1.1.0"
    mypy                  = "^0.931"
    numpy                 = "^1.22.1"
    opencv-contrib-python = "^4.5.5.62"
    pandas                = "^1.4.1"
    pathvalidate          = "^2.5.0"
    pre-commit            = "^2.17.0"
    prettytable           = "^3.2.0"
    pylint                = "^2.12.2"
    python                = ">=3.8,<3.11"
    regex                 = "^2022.3.15"
    spyder-kernels        = "^2.1.3"
    termcolor             = "^1.1.0"
    types-termcolor       = "^1.1.3"
    wheel                 = "^0.37.0"
[tool.poetry.scripts]
    LinkArtists = "ResizeImage.parser:parse_sub"
    ResizeImage = "ResizeImage.parser:parse_main"
[tool.black]
    line-length = 78
[tool.pycln]
    all = true
[tool.pyright]
    reportUnnecessaryIsInstance = false
